chore(cleanup + write up): moving functions for organisation; writing up tasks
